Native paragraph is working, but not yet commonized Moved native to skiko 0.0.0-SNAPSHOT-native9

+++ compose/ui/ui-text/src/desktopMain/kotlin/androidx/compose/ui/text/platform/DesktopParagraph.desktop.kt
@@ -550,7 +550,7 @@ internal class ParagraphBuilder(
         val ps = textStyleToParagraphStyle(textStyle)
 
         if (maxLines != Int.MAX_VALUE) {
-            ps.maxLinesCount = maxLines.toLong()
+            ps.maxLinesCount = maxLines
             ps.ellipsis = ellipsis
         }
